rename __class_getitem__ to from_callable
